contoller class명 수정
